Rewrite create test user function to use new user model.
